If at first you don't succeed, try try again

+++ Gateway/MinistryPlatform.Models/Group_Signup_Relationships.cs
@@ -1,12 +1,11 @@
 ﻿using System;
-using System.Data.SqlTypes;
 
 namespace MinistryPlatform.Models
 {
     public class GroupSignupRelationships
     {
         public int RelationshipId { get; set; }
-        public Byte RelationshipMinAge { get; set; }
-        public Byte RelationshipMaxAge { get; set; }
+        public int RelationshipMinAge { get; set; }
+        public int RelationshipMaxAge { get; set; }
     }
 }
